made changes to the wrong __init__ file
